+++ PSakefile.ps1
@@ -1,7 +1,8 @@
-
 Properties {
     Assert ($verbosity -ne $null) "`$verbosity should not be null. E.g. run with -Parameters @{ 'verbosity' = 'minimal'; }"
     
+    $WorkRoot = Get-RepositoryPath 'work' 
+
     # Console parameters.
     $BuildArgs = "/verbosity:$verbosity", '/maxcpucount', '/nodeReuse:false'
 
@@ -13,9 +14,10 @@ Properties {
     # Packaging properties:
     # - Release configuration
     # - Generate assembly versions (necessary for NuGet packaging)
+    # - Set the git commit hash
     # - Sign assemblies
-    # - Unconditionally hide internals (implies no white-box testing)
     # - Do not skip the generation of the Code Contracts reference assembly
+    # - Unconditionally hide internals (implies no white-box testing)
     $PackagingProps = `
         '/p:Configuration=Release',
         '/p:BuildGeneratedVersion=true',
@@ -26,13 +28,15 @@ Properties {
 
     # Default CI properties:
     # - Release configuration
-    # - Do not generate assembly versions
+    # - Do not generate assembly versions (=> no need to set the git commit hash)
     # - Do not sign assemblies
+    # - Do not skip the generation of the Code Contracts reference assembly
     # - Leak internals to enable all white-box tests.
     $CIProps = `
         '/p:Configuration=Release',
         '/p:BuildGeneratedVersion=false',
         '/p:SignAssembly=false',
+        '/p:SkipCodeContractsReferenceAssembly=false',
         '/p:VisibleInternals=true'
 
     # For static analysis, we hide internals, otherwise we might not truly 
@@ -58,9 +62,10 @@ TaskTearDown {
 
 Task Default -depends FastBuild
 
-# ==============================================================================
+# ------------------------------------------------------------------------------
+# Continuous Integration and development tasks
+# ------------------------------------------------------------------------------
 
-# Continuous Integration and development tasks:
 # - FastBuild, build then run Xunit tests in Debug configuration
 # - CI, default CI build
 # - Mock-Retail, mimic the Retail task
@@ -112,7 +117,9 @@ Task SecurityAnalysis {
         '/p:SkipDocumentation=true'
 }
 
-# ==============================================================================
+# ------------------------------------------------------------------------------
+# Retail tasks
+# ------------------------------------------------------------------------------
 
 # Retail tasks:
 # - Retail, package all
@@ -143,14 +150,47 @@ Task Retail-Build {
         '/p:Filter=_Build_'
 }
 
-# ==============================================================================
+# ------------------------------------------------------------------------------
+# MyGet tasks
+# ------------------------------------------------------------------------------
+
+Task Clean-MyGet {
+    $script:MyGetDir = Join-Path $WorkRoot 'myget'
+    $script:MyGetPkg = Join-Path $WorkRoot 'myget.7z'
+
+    if (Test-Path $MyGetDir) {
+        Remove-Item $MyGetDir -Force -Recurse
+    }
+    if (Test-Path $MyGetPkg) {
+        Remove-Item $MyGetPkg -Force
+    }
+}
+
+Task Build-MyGet {
+    # Force the value of "VisualStudioVersion", otherwise MSBuild won't publish the project on build.
+    # Cf. http://sedodream.com/2012/08/19/VisualStudioProjectCompatabilityAndVisualStudioVersion.aspx.
+    MSBuild $BuildArgs `
+        (Get-RepositoryPath 'tools', 'MyGet', 'MyGet.csproj') `
+        '/t:Clean;Build',
+        '/p:Configuration=Release;PublishProfile=NarvaloOrg;DeployOnBuild=true;VisualStudioVersion=12.0'
+}
+
+Task Publish-MyGet -depends Clean-MyGet, Build-MyGet {
+    . (Install-7Zip) -mx9 a $script:MyGetPkg $script:MyGetDir | Out-Null
+
+    Write-Host "A ready to publish package for MyGet may be found here: '$script:MyGetPkg'." -ForegroundColor Green
+}
+
+Task MyGet -depends Publish-MyGet
+
+# ------------------------------------------------------------------------------
+# Miscs
+# ------------------------------------------------------------------------------
 
 # Delete work directory.
 Task FullClean {
-    $workDir = Get-RepositoryPath 'work'
-
-    if (Test-Path $workDir) {
-        Remove-Item $workDir -Force -Recurse
+    if (Test-Path $WorkRoot) {
+        Remove-Item $WorkRoot -Force -Recurse
     }
 }